+++ src/main/kotlin/io/github/dlachouette/teamcity/github/web/AdminConsolePage.kt
@@ -77,6 +77,7 @@ class AdminConsolePage(
         model["set_testStats"] = serverSettings.checkRunTestStatsEnabled()
         model["set_timings"] = serverSettings.checkRunTimingsEnabled()
         model["set_infraNeutral"] = serverSettings.infraFailureNeutralEnabled()
+        model["set_cancelObsolete"] = serverSettings.cancelObsoleteEnabled()
         model["set_queueCleanup"] = serverSettings.queueCleanupEnabled()
         model["set_prTag"] = serverSettings.prTagEnabled()
         model["set_prTagPrefix"] = serverSettings.prTagPrefix()